"package 경로 변경, 구조화"
