+++ app/src/main/java/com/example/umc/Main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.umc
+package com.example.umc.Main
 
 import android.os.Bundle
 import android.view.View
@@ -6,6 +6,10 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import com.example.umc.Diet.DietDetailFragment
 import com.example.umc.Diet.HomeContainerFragment
+import com.example.umc.Mypage.MyFragment
+import com.example.umc.Quote.PriceFragment
+import com.example.umc.R
+import com.example.umc.Subscribe.SubFragment
 import com.example.umc.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
