+++ app/src/main/java/com/example/umc/Onboarding/MyActivity.kt
@@ -11,7 +11,6 @@ import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.example.umc.R
-import com.example.umc.Subscribe.SubscribeActivity
 
 class MyActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
