+++ app/src/main/java/com/example/umc/Onboarding/PriceActivity.kt
@@ -11,9 +11,7 @@ import android.widget.ImageButton
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
-import com.example.umc.Diet.DietActivity
 import com.example.umc.R
-import com.example.umc.Subscribe.SubscribeActivity
 
 class PriceActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
